pyproject.toml
[project]
name = "trevanquant"
version = "0.1.0"
description = "基于akshare的A股量化复盘系统"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "trevan", email = "trevan@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
keywords = ["quant", "stock", "analysis", "akshare", "trading"]

dependencies = [
    "akshare>=1.12.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "yagmail>=0.15.0",
    "jinja2>=3.1.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "loguru>=0.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trevanquant"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
